refactor(web): lift the retry-after parser and the wire out of the contact form
Both are about to have a second caller. The sign-in form posts to a
rate-limited API route just as the contact form does, so it needs the same
Retry-After parsing and the same words for a wait; and the animated wire that
shows a message going out to the API is the device that will tie the console
to the public site.

No behaviour change. formatWait moves verbatim rather than being rewritten,
so the contact form's copy is untouched.
